camera: Replace libcamera_client with android.hardware.camera.common@1.0-helper
libcamera_client is not a VNDK library and is not available to vendors in a Treblized system.
This camera HAL uses the library due to the CameraParamters class.
It is now available via a static VNDK library, android.hardware.camera.common@1.0-helper, which is available to vendors.
Ergo, we are switching to the library.